pyproject.toml
[project]
name = "git-update-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.46",
    "textual>=7.3.0",
    "textual-dev>=1.8.0",
]
